refactor: moved project managment into its own component
